+++ database/migrations/2018_10_01_140759_create_employee_school_table.php
@@ -14,9 +14,9 @@ class CreateEmployeeSchoolTable extends Migration
     public function up()
     {
         Schema::create('employee_school', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('employee_id');
-            $table->unsignedInteger('school_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('employee_id');
+            $table->unsignedBigInteger('school_id');
             $table->timestamp('start_at')->nullable();
             $table->timestamp('graduated_at')->nullable();
             $table->timestamps();
